- Remove stacked_locater from stacked_package - Rename router to router_annotation and fix imports
