+++ packages/stacked/lib/src/code_generation/router_annotation/parameters.dart
