+++ packages/stacked/lib/src/code_generation/router_annotation/route_guard.dart
