+++ packages/stacked/lib/src/code_generation/router_annotation/router_utils.dart
