+++ packages/stacked/lib/src/code_generation/router_annotation/stacked_route_wrapper.dart
