+++ packages/stacked/lib/src/code_generation/router_annotation/transitions_builders.dart
