+++ packages/stacked/lib/src/code_generation/router_annotation/uri_extension.dart
