+++ packages/stacked/lib/stacked.dart
@@ -9,13 +9,12 @@ export 'src/state_management/view_model_builder_widget.dart';
 export 'src/state_management/view_model_widget.dart';
 export 'src/state_management/skeleton_loader.dart';
 
-export 'src/code_generation/router/extended_navigator.dart';
-export 'src/code_generation/router/router_base.dart';
-export 'src/code_generation/router/route_guard.dart';
-export 'src/code_generation/router/route_def.dart';
-export 'src/code_generation/router/router_utils.dart';
-export 'src/code_generation/router/transitions_builders.dart';
-export 'src/code_generation/stacked_locator.dart';
+export 'src/code_generation/router_annotation/extended_navigator.dart';
+export 'src/code_generation/router_annotation/router_base.dart';
+export 'src/code_generation/router_annotation/route_guard.dart';
+export 'src/code_generation/router_annotation/route_def.dart';
+export 'src/code_generation/router_annotation/router_utils.dart';
+export 'src/code_generation/router_annotation/transitions_builders.dart';
 
 export 'src/reactive/reactive_value/reactive_value.dart';
 export 'src/reactive/reactive_list/reactive_list.dart';
